+++ admin/src/components/PanelV2.js
@@ -2,20 +2,22 @@ import { Dialog, Transition } from "@headlessui/react";
 import React, { Fragment } from "react";
 import CloseSvg from "../assets/Close";
 
-export default function PanelV2({ children, title, open, onClose }) {
+export default function PanelV2({ children, title, open, onClose, size = "lg", darkBg = true }) {
   return (
     <Transition.Root show={open} as={Fragment}>
       <Dialog as="div" className="relative z-50" onClose={onClose}>
-        <Transition.Child
-          as={Fragment}
-          enter="ease-in-out duration-500"
-          enterFrom="opacity-0"
-          enterTo="opacity-100"
-          leave="ease-in-out duration-500"
-          leaveFrom="opacity-100"
-          leaveTo="opacity-0">
-          <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
-        </Transition.Child>
+        {darkBg && (
+          <Transition.Child
+            as={Fragment}
+            enter="ease-in-out duration-500"
+            enterFrom="opacity-0"
+            enterTo="opacity-100"
+            leave="ease-in-out duration-500"
+            leaveFrom="opacity-100"
+            leaveTo="opacity-0">
+            <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
+          </Transition.Child>
+        )}
 
         <div className="fixed inset-0 overflow-hidden">
           <div className="absolute inset-0 overflow-hidden">
@@ -28,7 +30,7 @@ export default function PanelV2({ children, title, open, onClose }) {
                 leave="transform transition ease-in-out duration-500 sm:duration-700"
                 leaveFrom="translate-x-0"
                 leaveTo="translate-x-full">
-                <Dialog.Panel className="pointer-events-auto w-screen max-w-lg">
+                <Dialog.Panel className={`pointer-events-auto w-screen max-w-${size}`}>
                   <div className="flex h-full flex-col overflow-y-auto bg-white pt-2 shadow-xl">
                     <div className="px-4 sm:px-6">
                       <div className="flex items-start justify-between">